+++ package.json
@@ -23,6 +23,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.4",
+    "react-intersection-observer": "^9.16.0",
     "zustand": "^5.0.4"
   },
   "devDependencies": {
